+++ CHANGELOG.md
@@ -18,7 +18,9 @@ All notable changes to this project will be documented in this file. This change
 - Baseline AgentLang version `0.7.0-alpha1`
 - Fix new-project template
   - [Todo] CamelCase project names expanded to component directory name
-  - [Todo] DesignStudio-friendly quoted form in generated resolver project
+  - DesignStudio-friendly quoted form in generated project
+- Add commands
+  - [Todo] `lint` - Run a linter (EDN-check all `.al` files, and more)
 
 ## [0.6.2] - 2025-03-18