+++ CHANGELOG.md
@@ -17,7 +17,7 @@ All notable changes to this project will be documented in this file. This change
 
 - Baseline AgentLang version `0.7.0-alpha1`
 - Fix new-project template
-  - [Todo] CamelCase project names expanded to component directory name
+  - CamelCase project names expanded to component directory name
   - DesignStudio-friendly quoted form in generated project
 - Add commands
   - [Todo] `lint` - Run a linter (EDN-check all `.al` files, and more)